+++ src/pages/AdminPage.js
@@ -6,7 +6,7 @@ import { Link } from "react-router-dom";
 import { Room } from "../models/Room";
 import { Task, TaskList } from "../models/TaskList.js";
 import background from "../images/stars-background.jpg";
-import FrontendTaskList from "../components/TaskListComponent.js";
+import FrontendTaskList from "../test-components/TaskListComponent.js";
 import { useNavigate } from 'react-router-dom';
 
 let room = null;
